+++ lib/subspace/commands/ansible.rb
@@ -3,7 +3,7 @@ module Subspace
     module Ansible
       def ansible_command(command, *args)
         update_ansible_cfg
-        Dir.chdir "config/provision" do
+        Dir.chdir "config/subspace" do
           say ">> Running #{command} #{args.join(' ')}"
           system(command, *args, out: $stdout, err: $stderr)
           say "<< Done"