+++ lib/subspace/commands/configure.rb
@@ -16,12 +16,12 @@ class Subspace::Commands::Configure < Subspace::Commands::Base
   private
 
   def update_host_configuration(host)
-    say "Generating config/provisiong/host_vars/#{host}"
+    say "Generating config/subspace/host_vars/#{host}"
     template "host_vars/template", "host_vars/#{host}", Subspace.config.binding_for(host: host)
   end
 
   def update_group_configuration(group)
-    say "Generating config/provisiong/group_vars/#{group}"
+    say "Generating config/subspace/group_vars/#{group}"
     template "group_vars/template", "group_vars/#{group}", Subspace.config.binding_for(group: group)
   end
 end